uint256 -> uint

+++ contracts/BinaryOption.sol
@@ -16,18 +16,18 @@ contract BinaryOption {
 
     BinaryOptionMarket public market;
 
-    mapping(address => uint256) public bidOf;
-    uint256 public totalBids;
+    mapping(address => uint) public bidOf;
+    uint public totalBids;
 
-    mapping(address => uint256) public balanceOf;
-    uint256 public totalSupply;
+    mapping(address => uint) public balanceOf;
+    uint public totalSupply;
 
     // The argument order is allowance[owner][spender]
-    mapping(address => mapping(address => uint256)) public allowance;
+    mapping(address => mapping(address => uint)) public allowance;
 
     /* ========== CONSTRUCTOR ========== */
 
-    constructor(address initialBidder, uint256 initialBid) public {
+    constructor(address initialBidder, uint initialBid) public {
         market = BinaryOptionMarket(msg.sender);
         bidOf[initialBidder] = initialBid;
         totalBids = initialBid;
@@ -35,40 +35,40 @@ contract BinaryOption {
 
     /* ========== VIEWS ========== */
 
-    function price() public view returns (uint256) {
+    function price() public view returns (uint) {
         return market.senderPrice();
     }
 
-    function claimableBy(address account) public view returns (uint256) {
+    function claimableBy(address account) public view returns (uint) {
         return bidOf[account].divideDecimal(price());
     }
 
-    function totalClaimable() public view returns (uint256) {
+    function totalClaimable() public view returns (uint) {
         return totalBids.divideDecimal(price());
     }
 
-    function totalExercisable() external view returns (uint256) {
+    function totalExercisable() external view returns (uint) {
         return totalSupply + totalClaimable();
     }
 
     /* ========== MUTATIVE FUNCTIONS ========== */
 
     // This must only be invoked during bidding.
-    function bid(address bidder, uint256 newBid) external onlyMarket {
+    function bid(address bidder, uint newBid) external onlyMarket {
         bidOf[bidder] = bidOf[bidder].add(newBid);
         totalBids = totalBids.add(newBid);
     }
 
     // This must only be invoked during bidding.
-    function refund(address bidder, uint256 newRefund) external onlyMarket {
+    function refund(address bidder, uint newRefund) external onlyMarket {
         // The safe subtraction will catch refunds that are too large.
         bidOf[bidder] = bidOf[bidder].sub(newRefund);
         totalBids = totalBids.sub(newRefund);
     }
 
     // This must only be invoked after bidding.
-    function claim(address claimant) external onlyMarket returns (uint256 optionsClaimed) {
-        uint256 claimable = claimableBy(claimant);
+    function claim(address claimant) external onlyMarket returns (uint optionsClaimed) {
+        uint claimable = claimableBy(claimant);
         // No options to claim? Nothing happens.
         if (claimable == 0) {
             return 0;
@@ -88,7 +88,7 @@ contract BinaryOption {
 
     // This must only be invoked after maturity.
     function exercise(address claimant) external onlyMarket {
-        uint256 balance = balanceOf[claimant];
+        uint balance = balanceOf[claimant];
 
         if (balance == 0) {
             return;
@@ -112,10 +112,10 @@ contract BinaryOption {
     // This should only operate after bidding;
     // Since options can't be claimed until after bidding, all balances are zero until that time.
     // So we don't need to explicitly check the timestamp to prevent transfers.
-    function internalTransfer(address _from, address _to, uint256 _value) internal returns (bool success) {
+    function internalTransfer(address _from, address _to, uint _value) internal returns (bool success) {
         require(_to != address(0) && _to != address(this), "Cannot transfer to this address.");
 
-        uint256 fromBalance = balanceOf[_from];
+        uint fromBalance = balanceOf[_from];
         require(_value <= fromBalance, "Insufficient balance.");
 
         balanceOf[_from] = fromBalance.sub(_value);
@@ -125,19 +125,19 @@ contract BinaryOption {
         return true;
     }
 
-    function transfer(address _to, uint256 _value) public returns (bool success) {
+    function transfer(address _to, uint _value) public returns (bool success) {
         return internalTransfer(msg.sender, _to, _value);
     }
 
-    function transferFrom(address _from, address _to, uint256 _value) public returns (bool success) {
-        uint256 fromAllowance = allowance[_from][msg.sender];
+    function transferFrom(address _from, address _to, uint _value) public returns (bool success) {
+        uint fromAllowance = allowance[_from][msg.sender];
         require(_value <= fromAllowance, "Insufficient allowance.");
 
         allowance[_from][msg.sender] = fromAllowance.sub(_value);
         return internalTransfer(_from, _to, _value);
     }
 
-    function approve(address _spender, uint256 _value) public returns (bool success) {
+    function approve(address _spender, uint _value) public returns (bool success) {
         require(_spender != address(0));
         allowance[msg.sender][_spender] = _value;
         emit Approval(msg.sender, _spender, _value);
@@ -155,6 +155,6 @@ contract BinaryOption {
 
     event Issued(address indexed account, uint value);
     event Burned(address indexed account, uint value);
-    event Transfer(address indexed from, address indexed to, uint256 value);
-    event Approval(address indexed owner, address indexed spender, uint256 value);
+    event Transfer(address indexed from, address indexed to, uint value);
+    event Approval(address indexed owner, address indexed spender, uint value);
 }
