+++ contracts/interfaces/IBinaryOption.sol
@@ -5,13 +5,13 @@ import "../interfaces/IERC20.sol";
 
 contract IBinaryOption is IERC20 {
     IBinaryOptionMarket public market;
-    mapping(address => uint256) public bidOf;
-    uint256 public totalBids;
+    mapping(address => uint) public bidOf;
+    uint public totalBids;
 
-    function price() public view returns (uint256);
-    function claimableBy(address account) public view returns (uint256);
-    function totalClaimable() public view returns (uint256);
-    function totalExercisable() external view returns (uint256);
+    function price() public view returns (uint);
+    function claimableBy(address account) public view returns (uint);
+    function totalClaimable() public view returns (uint);
+    function totalExercisable() external view returns (uint);
 
     event Issued(address indexed account, uint value);
     event Burned(address indexed account, uint value);
